requirements.txt
pytorch==1.3.1
transformers==2.8.0
 
